+++ package.json
@@ -9,8 +9,8 @@
     "format-check": "prettier --check '**/*.js'",
     "lint": "eslint src/**/*.js",
     "package": "NODE_OPTIONS=--openssl-legacy-provider ncc build src/main.js -o dist --source-map --license licenses.txt",
-    "test": "jest",
-    "ci-test": "jest",
+    "test": "GIT_DIFF_JEST_TEST=true jest",
+    "ci-test": "GIT_DIFF_JEST_TEST=true jest",
     "all": "npm run format && npm run lint && npm run package",
     "bundle": "npm run format && npm run package"
   },
