+++ src/main.js
@@ -4,4 +4,7 @@ export async function run() {
   await gitDiff()
 }
 
-run()
+/* istanbul ignore next */
+if (process.env.GIT_DIFF_JEST_TEST !== 'true') {
+  run()
+}
